MOD IHM : Add a light mode.
